django/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
deprecation==2.1.0
Django==4.2.5
django-cors-headers==3.7.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
djoser==2.0.5
django-jazzmin==2.4.8
drf-yasg==1.20.0
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
josepy==1.14.0
mozilla-django-oidc==3.0.0
packaging==23.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.7.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==5.3.0